backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.1
scikit-learn==1.3.1
numpy==1.26.0
Pillow==10.1.0
dj-rest-auth==5.0.1
django-allauth==0.57.0
gunicorn==21.2.0
whitenoise==6.6.0